+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- 🪟 windows-legacy build now works on windows 7 without VxKex!
+
 ## [v5.1.7] - 2026-02-12
 
 ### Fixed